Extract josh-github-keyring from josh-cli
Move credential storage (FileCredentialStore, default_store,
StoredToken, load_stored_token) into a dedicated forge crate
so the merge queue can reuse it to obtain GitHub credentials
without depending on josh-cli.

josh-cli now depends on josh-github-keyring instead of the
raw keyring crate directly.
