+++ forges/josh-github-keyring/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "josh-github-keyring"
+version = "26.5.8"
+edition = "2024"
+authors = ["Josh Project authors <contact@josh-project.dev>"]
+description = "Credential storage for Josh GitHub integration"
+license-file = "../../LICENSE"
+repository = "https://github.com/josh-project/josh"
+keywords = ["git", "github", "monorepo", "credential"]
+
+[dependencies]
+anyhow.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+chrono = { workspace = true, features = ["clock"] }
+dirs.workspace = true
+keyring = { version = "^3.6", features = ["apple-native"] }
+
+josh-github-auth.workspace = true
+
+[features]
+codesign = []
