+++ josh-cli/Cargo.toml
@@ -26,13 +26,13 @@ dirs.workspace = true
 reqwest.workspace = true
 reqwest-middleware.workspace = true
 
-keyring = { version = "^3.6", features = ["apple-native"] }
 arboard = { version = "^3.6", default-features = false }
 
 josh-changes.workspace = true
 josh-core.workspace = true
 josh-compose.workspace = true
 josh-github-auth.workspace = true
+josh-github-keyring.workspace = true
 josh-github-graphql.workspace = true
 josh-github-changes.workspace = true
 josh-graphql.workspace = true
