+++ josh-cli/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod commands;
 pub mod config;
 pub mod forge;
-pub mod keyring;
 pub mod remote_ops;
